Test: User authorization logic test - TODO: complete authorization in - BadgeController.java, etc... - CategoryServiceImpl.java, etc...

+++ src/main/java/com/oss/gitborad/service/impl/BookmarkServiceImpl.java
@@ -64,7 +64,11 @@ public class BookmarkServiceImpl implements BookmarkService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, Long userId) {
+        User user = userRepository.findById(userId).orElse(null);
+        assert user != null;
+        if(!user.getId().equals(userId))
+            return;
         bookmarkRepository.deleteById(id);
     }
 }
